Add basic smoke test for UprobeStats that can run in CI
This change adds a test for UprobeStats that we can run in presubmit.
Now that UprobeStats can write to StatsD, we can confirm that triggering
a probe and interacting with the device in the expected way leads to the
expected atom values being present in StatsD.

Also adds java and Android.bp formatting to `PREUPLOAD.cfg`.

This iteration of the test does not work on x86, and thus does not
enable the test yet in presubmit. This will be added with a later
change.

Bug: 296108553
Test: atest uprobestats-test
Flag: NONE UprobeStats does not yet run on devices

+++ src/Bpf.cpp
@@ -90,7 +90,7 @@ std::vector<T> pollRingBuf(const char *mapPath, int timeoutMs) {
   return vec;
 }
 
-template std::vector<uint32_t> pollRingBuf(const char *mapPath, int timeoutMs);
+template std::vector<uint64_t> pollRingBuf(const char *mapPath, int timeoutMs);
 template std::vector<CallResult> pollRingBuf(const char *mapPath,
                                              int timeoutMs);
